further reorg of sparse routines
