+++ pandas/tests/indexes/datetimes/test_date_range.py
@@ -10,7 +10,6 @@ from pandas.tests.series.common import TestData
 class TestTimeSeries(TestData, tm.TestCase):
     _multiprocess_can_split_ = True
 
-
     def test_date_range_gen_error(self):
         rng = date_range('1/1/2000 00:00', '1/1/2000 00:18', freq='5min')
         self.assertEqual(len(rng), 4)
